feat(profile-post-popup): extract post-popup and add delete button
