+++ lib/views/pages/profile/posts_info/popup/post_popup.dart
@@ -1,0 +1,26 @@
+import "package:flutter/material.dart";
+
+class PostPopUp extends StatelessWidget {
+  const PostPopUp({
+    super.key,
+    required this.title,
+    required this.description,
+  });
+
+  final String title;
+  final String description;
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      title: Text(title),
+      content: Text(description),
+      actions: <Widget>[
+        IconButton(
+          onPressed: () {},
+          icon: const Icon(Icons.delete),
+        ),
+      ],
+    );
+  }
+}
